Small clarification on example
This is a small clarification for those (like me) who might have read the example a bit too fast and were confused by the `undefined` output in the console. Instead, the code snippet just runs `esbuild.build` and the text emphasizes that the expected output lives in `dist/bytes.esm.js`.